global: documents module with simple minter/fetcher
