+++ invenio_app_ils/circulation/stats/views.py
@@ -17,13 +17,10 @@ from invenio_records_rest.utils import obj_or_import_string
 from invenio_rest import ContentNegotiatedMethodView
 
 from invenio_app_ils.circulation.stats.api import fetch_most_loaned_documents
+from invenio_app_ils.documents.api import DOCUMENT_PID_FETCHER, \
+    DOCUMENT_PID_TYPE
 from invenio_app_ils.errors import InvalidParameterError
 from invenio_app_ils.permissions import need_permissions
-
-from invenio_app_ils.pidstore.pids import (  # isort:skip
-    DOCUMENT_PID_FETCHER,
-    DOCUMENT_PID_TYPE,
-)
 
 
 def create_most_loaned_documents_view(blueprint, app):
