+++ invenio_app_ils/document_requests/indexer.py
@@ -13,9 +13,9 @@ from celery import shared_task
 from flask import current_app
 from invenio_indexer.api import RecordIndexer
 
+from invenio_app_ils.documents.api import DOCUMENT_PID_TYPE
 from invenio_app_ils.indexer import ReferencedRecordsIndexer
-from invenio_app_ils.pidstore.pids import DOCUMENT_PID_TYPE, \
-    DOCUMENT_REQUEST_PID_TYPE
+from invenio_app_ils.pidstore.pids import DOCUMENT_REQUEST_PID_TYPE
 from invenio_app_ils.proxies import current_app_ils
