+++ invenio_app_ils/documents/jsonresolvers/document_item.py
