+++ invenio_app_ils/documents/jsonresolvers/document_relations.py
@@ -10,8 +10,8 @@
 import jsonresolver
 from werkzeug.routing import Rule
 
+from invenio_app_ils.documents.api import Document
 from invenio_app_ils.jsonresolver.api import get_pid_or_default
-from invenio_app_ils.records.api import Document
 
 # Note: there must be only one resolver per file,
 # otherwise only the last one is registered
