+++ invenio_app_ils/documents/jsonresolvers/document_stock.py
