+++ invenio_app_ils/documents/mappings/v6/documents/document-v1.0.0.json
