+++ invenio_app_ils/documents/mappings/v7/documents/document-v1.0.0.json
