+++ invenio_app_ils/documents/schemas/documents/document-v1.0.0.json
