+++ invenio_app_ils/eitems/indexer.py
@@ -13,8 +13,9 @@ from celery import shared_task
 from flask import current_app
 from invenio_indexer.api import RecordIndexer
 
+from invenio_app_ils.documents.api import DOCUMENT_PID_TYPE
 from invenio_app_ils.indexer import ReferencedRecordsIndexer
-from invenio_app_ils.pidstore.pids import DOCUMENT_PID_TYPE, EITEM_PID_TYPE
+from invenio_app_ils.pidstore.pids import EITEM_PID_TYPE
 from invenio_app_ils.proxies import current_app_ils
