+++ invenio_app_ils/records/resolver/jsonresolver/item_document.py
@@ -10,11 +10,12 @@
 import jsonresolver
 from werkzeug.routing import Rule
 
+from invenio_app_ils.documents.api import Document
 from invenio_app_ils.jsonresolver.api import \
     get_field_value_for_record as get_field_value
 from invenio_app_ils.jsonresolver.api import get_pid_or_default
 
-from ...api import Document, Item
+from ...api import Item
 
 # Note: there must be only one resolver per file,
 # otherwise only the last one is registered
